Complete Step 6 + some debugging
Need to go double-check some more stuff. Debugged a major issue with the removeLibrary function;
if a book that is not the first in the library is removed, it would cause an issue where even after
removing every consequent book, the books would still be displayed and not fully removed in the myLibrary array.

Taking a break! Good progress